requirements.txt
customtkinter
Pillow
tiktoken
tkinterdnd2
python-docx
pypdf
